First Push
Started from the beginning of the Buildspace project and got through to "Add in your prompt" – which is the last section (3/3) in the module entitled "Set up your own gpt-3 writer"

Comments on course: It's interesting to get thrown into the sharks, but I need a better understanding of the depths of JavaScript: Node, React, and Next to be precise. And then of course there is the MERN stack (which adds MongoDB to the mix).

I will continue with Scrimba and ChatGPT and mentors/peers/mentees.